requirements.txt
Flask==3.0.2
google-cloud-storage==2.14.0
google-cloud-bigquery==3.17.2
Werkzeug==3.0.1
gunicorn==21.2.0
